adds

+++ project/src/main/java/com/emranhss/project/dto/DistrictResponseDTO.java
@@ -5,21 +5,25 @@ import java.util.List;
 public class DistrictResponseDTO {
 
 
-    private int id;
+    private long id;
     private String name;
     private List<Integer> policeStations;
 
     // Getters and setters
-    public int getId() {
+
+
+    public long getId() {
         return id;
     }
-    public void setId(int id) {
+
+    public void setId(long id) {
         this.id = id;
     }
 
     public String getName() {
         return name;
     }
+
     public void setName(String name) {
         this.name = name;
     }
@@ -27,21 +31,8 @@ public class DistrictResponseDTO {
     public List<Integer> getPoliceStations() {
         return policeStations;
     }
+
     public void setPoliceStations(List<Integer> policeStations) {
         this.policeStations = policeStations;
     }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 }
