+++ project/src/main/java/com/emranhss/project/repository/IDivisionRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IDivisionRepo extends JpaRepository<Division,Long> {
+public interface IDivisionRepo extends JpaRepository<Division,Integer> {
 }
